cambios en el tipo de variables

+++ TSSDistribiciones/src/distribuciones/Media.java
@@ -5,6 +5,7 @@
  */
 package distribuciones;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -12,14 +13,15 @@ import java.util.List;
  * @author Douglas
  */
 public class Media {
-    private final List<Float> datos;
-
-    public Media(List<Float> datos) {
+    private final List<Double> datos;
+    
+    public Media(List<Double> datos) {
         this.datos = datos;
     }
 
-    public float calculoMedia(){
-        float sumatoria = 0f;
+    public double calculoMedia(){
+
+        double sumatoria = 0;
         sumatoria = datos.stream().map((dato) -> dato).reduce(sumatoria, (accumulator, _item) -> accumulator + _item);
         return sumatoria /datos.size();
     }
